+++ middleware/routers/wrapper_info.go
@@ -18,7 +18,6 @@ func NewInfoRouter(conf fastapi.Config, prefix ...string) fastapi.GroupRouter {
 		Title:   conf.Title,
 		Version: conf.Version,
 		Desc:    conf.Description,
-		Debug:   conf.Debug,
 	}
 	if len(prefix) > 0 {
 		r.prefix = prefix[0]
@@ -33,7 +32,6 @@ type WrapperInfoRouter struct {
 	Title   string
 	Version string
 	Desc    string
-	Debug   bool
 	Tag     []string
 }
 
@@ -52,13 +50,8 @@ func (r *WrapperInfoRouter) Summary() map[string]string {
 		"GetTitle":       "获取软件名",
 		"GetDescription": "获取软件描述信息",
 		"GetVersion":     "获取软件版本号",
-		"GetDebug":       "获取调试开关",
 		"GetHeartbeat":   "心跳检测",
 	}
-}
-
-func (r *WrapperInfoRouter) Description() map[string]string {
-	return map[string]string{}
 }
 
 func (r *WrapperInfoRouter) GetTitle(c *fastapi.Context) (string, error) {
@@ -71,10 +64,6 @@ func (r *WrapperInfoRouter) GetDescription(c *fastapi.Context) (string, error) {
 
 func (r *WrapperInfoRouter) GetVersion(c *fastapi.Context) (string, error) {
 	return r.Version, nil
-}
-
-func (r *WrapperInfoRouter) GetDebug(c *fastapi.Context) (bool, error) {
-	return r.Debug, nil
 }
 
 func (r *WrapperInfoRouter) GetHeartbeat(c *fastapi.Context) (string, error) {