+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # CHANGELOG
 
+## 0.2.8 - (2024-04-22)
+
+### Feat
+
+- expose MuxContext;
+- 不再打算支持泛型路由;
+- ~~(2024-04-26) 路由组新增`ErrorFormatter`方法~~;
+- (2024-04-26) 支持设置全局错误处理函数`RouteErrorFormatter`及其数据模型，并为其生成文档;
+
+### Fix
+
+- (2024-04-27) 依赖函数返回错误时同样通过`RouteErrorFormatter`进行格式化
+
+### BREAKING
+
+- (2024-04-26) 移除`Wrapper.Debug`参数;
+
 ## 0.2.7 - (2024-04-19)
 
 ### Feat